:sparkles: Separate file for FileFinder
